roachpb: fix coverage build
Moves all String tests to a new file that lines in roachpb_test so that
it can import keys and enjoy the magic of the pretty printer.

This fixes the coverage build which links things in an unusual way.
